+++ day15/part2/generators/main_test.go
@@ -8,10 +8,10 @@ import (
 
 var generatorNextTT = []struct {
 	name                             string
-	memA, memB                       int
-	factorA, factorB                 int
-	moduloCriteriaA, moduloCriteriaB int
-	expected                         int
+	memA, memB                       int64
+	factorA, factorB                 int64
+	moduloCriteriaA, moduloCriteriaB int64
+	expected                         int64
 }{
 	{"provided example", 65, 8921, 16807, 48271, 4, 8, 309},
 }
